[CF] adapt to a new version of Telegram API
- upgrade telegrambots spring boot starter major version (4 -> 5)
- int -> long for userId and some minor changes
- fix typo in docker-compose files

+++ src/main/java/ru/maza/telegram/controller/BotController.java
@@ -121,7 +121,7 @@ public class BotController extends TelegramLongPollingBot {
     @Override
     @SneakyThrows
     public void onUpdateReceived(Update update) {
-        Integer userId = telegramService.getUser(update).getId();
+        Long userId = telegramService.getUser(update).getId();
         UserDto userDto = textInfraService.saveUser(update, userId);
 
         if (update.hasMessage()) {
